Factor out morph_mode into individual classes
