+++ lib/stimulus_reflex/morph_mode.rb
@@ -1,0 +1,17 @@
+module StimulusReflex
+  class MorphMode
+    include StimulusReflex::Broadcaster
+
+    def page?
+      false
+    end
+
+    def nothing?
+      false
+    end
+
+    def selector?
+      false
+    end
+  end
+end
