+++ lib/stimulus_reflex/morph_mode/nothing_morph_mode.rb
@@ -1,0 +1,15 @@
+module StimulusReflex
+  class NothingMorphMode < MorphMode
+    def broadcast
+      broadcast_message subject: "nothing", data: data
+    end
+
+    def to_sym
+      :nothing
+    end
+
+    def nothing?
+      true
+    end
+  end
+end
