+++ lib/stimulus_reflex/morph_mode/selector_morph_mode.rb
@@ -1,0 +1,15 @@
+module StimulusReflex
+  class SelectorMorphMode < MorphMode
+    def broadcast
+      broadcast_message subject: "selector", data: data
+    end
+
+    def to_sym
+      :selector
+    end
+
+    def selector?
+      true
+    end
+  end
+end
